feat: update set of handle nft
