+++ src/app/admin/page.tsx
@@ -1,12 +1,14 @@
 import BicCard from "@/components/admin/BicCard";
 import HandleCard from "@/components/admin/HandleCard";
+import HandleControllerCard from "@/components/admin/HandleControllerCard";
 
 function Admin() {
 
     return (
         <>
             <BicCard/>
-            <HandleCard/>
+            <HandleCard />
+            <HandleControllerCard/>
         </>
     )
 }
