+++ src/components/admin/HandleCard.tsx
@@ -1,231 +1,127 @@
 "use client";
 
-import { useAccount, useWalletClient } from "wagmi";
-import { createPublicClient, getContract, http } from "viem";
-import { HandleControllerConfig } from "@/contract/contractConfig";
+import { useWalletClient } from "wagmi";
+import { createPublicClient, getAddress, getContract, http, zeroAddress } from "viem";
+import { HandleConfig } from "@/contract/contractConfig";
 import { useCallback, useEffect, useState } from "react";
+import { handleAddresses } from "@/util/constant";
 
-type HandleControllerConfig = {
-  marketplace: string;
-  verifier: string;
-  forwarder: string;
-  owner: string;
-  collector: string;
-  auctionConfig: {
-    buyoutBidAmount: string;
-    timeBufferInSeconds: string;
-    bidBufferBps: string;
-  };
-};
 export default function HandleCard() {
   const { data: walletClient } = useWalletClient();
-  const [config, setConfig] = useState<HandleControllerConfig>();
-  const [owner, setOwner] = useState<string>();
-  const [forwarder, setForwarder] = useState<string>();
-  const [verifier, setVerifier] = useState<string>();
-  const [marketplace, setMarketplace] = useState<string>();
-  const [collector, setCollector] = useState<string>();
-  const [auctionConfig, setAuctionConfig] = useState<{
-    buyoutBidAmount: string;
-    timeBufferInSeconds: string;
-    bidBufferBps: string;
-  }>({
-    bidBufferBps: "0",
-    timeBufferInSeconds: "0",
-    buyoutBidAmount: "0",
-  });
 
-  const handleControllerContract = useCallback(() => {
+  const [handleSelected, setHandleSelected] =
+    useState<`0x${string}`>(handleAddresses[0].address);
+  const [name, setName] = useState<string>();
+  const [symbol, setSymbol] = useState<string>();
+  const [controller, setController] = useState<string>();
+  const [handleTokenURI, setHandleTokenURI] = useState<string>();
+  const [operator, setOperator] = useState<string>();
+
+  const handleContract = useCallback(() => {
     if (!walletClient) return;
     const publicClient = createPublicClient({
       transport: http(),
       chain: walletClient.chain,
     });
-
+    const handleConfig = HandleConfig(handleSelected);
     const contract = getContract({
-      ...HandleControllerConfig,
+      ...handleConfig,
       client: {
         wallet: walletClient,
         public: publicClient,
       },
     });
     return contract;
-  }, [walletClient]);
+  }, [walletClient, handleSelected]);
 
   const fetchConfig = async () => {
     if (!walletClient) return;
-    const contract = handleControllerContract();
+    const contract = handleContract();
     if (!contract) return;
+    if(!handleSelected) return;
 
-    const [marketplace, verifier, owner, forwarder, collector, auctionConfig] =
-      await Promise.all([
-        contract.read.marketplace(),
-        contract.read.verifier(),
-        contract.read.owner(),
-        contract.read.forwarder(),
-        contract.read.collector(),
-        contract.read.auctionConfig(),
-      ]);
+    const [name, symbol, operator, controller, handleTokenURI] = await Promise.all([
+      contract.read.name(),
+      contract.read.symbol(),
+      contract.read.OPERATOR(),
+      contract.read.CONTROLLER(),
+      contract.read.getHandleTokenURIContract(),
+    ]);
 
-    setOwner(owner as string);
-    setForwarder(forwarder as string);
-    setVerifier(verifier as string);
-    setMarketplace(marketplace as string);
-    setCollector(collector as string);
-
-    setAuctionConfig({
-      buyoutBidAmount: (auctionConfig as any)[0]?.toString(),
-      timeBufferInSeconds: (auctionConfig as any)[1]?.toString(),
-      bidBufferBps: (auctionConfig as any)[2]?.toString(),
-    });
+    setName(name as string);
+    setSymbol(symbol as string);
+    setHandleTokenURI(handleTokenURI as string);
+    setController(controller as string);
+    setOperator(operator as string);
   };
 
   const handleError = (error: any) => {
     alert(error?.message || "Unknown error");
   };
 
-  const onUpdateOwner = async () => {
+  const onUpdateOperator = async () => {
     try {
-      const contract = handleControllerContract();
+      const contract = handleContract();
       if (!contract) {
         // Handle notification
         return;
       }
-      if (!owner) {
+      if (!operator) {
         // Handle notification
         return;
       }
-      const tx = await contract.write.transferOwnership([owner]);
+      const tx = await contract.write.setOperator([operator]);
 
       // Handle success
-      console.log("🚀 ~ onUpdateOwner ~ tx:", tx);
+      console.log("🚀 ~ onUpdateOperator ~ tx:", tx);
       fetchConfig();
-      alert("Owner updated successfully");
+      alert("Operator updated successfully");
     } catch (error) {
       // Handle error
-      console.log("🚀 ~ onUpdateOwner ~ error:", error);
+      console.log("🚀 ~ onUpdateOperator ~ error:", error);
     }
   };
 
-  const onUpdateCollector = async () => {
+  const onUpdateController = async () => {
     try {
-      const contract = handleControllerContract();
+      const contract = handleContract();
       if (!contract) {
         // Handle notification
         return;
       }
-      if (!owner) {
+      if (!controller) {
         // Handle notification
         return;
       }
-      const tx = await contract.write.setCollector([collector]);
+      const tx = await contract.write.setController([controller]);
 
       // Handle success
-      console.log("🚀 ~ onUpdateCollector ~ tx:", tx);
+      console.log("🚀 ~ onUpdateController ~ tx:", tx);
       fetchConfig();
-      alert("Collector updated successfully");
+      alert("Controllet updated successfully");
     } catch (error) {
       // Handle error
-      console.log("🚀 ~ onUpdateCollector ~ error:", error);
-      handleError(error);
+      console.log("🚀 ~ onUpdateController ~ error:", error);
     }
   };
 
-  const onUpdateAuctionConfig = async () => {
+  const onUpdateHandleTokenURI = async () => {
     try {
-      const contract = handleControllerContract();
+      const contract = handleContract();
       if (!contract) {
         // Handle notification
         return;
       }
-      if (!owner) {
+      if (!handleTokenURI) {
         // Handle notification
         return;
       }
-      const tx = await contract.write.setAuctionMarketplaceConfig([
-        [
-          auctionConfig?.buyoutBidAmount,
-          auctionConfig?.timeBufferInSeconds,
-          auctionConfig?.bidBufferBps,
-        ],
-      ]);
-
-      // Handle success
-      console.log("🚀 ~ onUpdateCollector ~ tx:", tx);
-      fetchConfig();
-      alert("Auction config updated successfully");
-    } catch (error) {
-      // Handle error
-      console.log("🚀 ~ onUpdateCollector ~ error:", error);
-      handleError(error);
-    }
-  };
-
-  const onUpdateMarketplace = async () => {
-    try {
-      const contract = handleControllerContract();
-      if (!contract) {
-        // Handle notification
-        return;
-      }
-      if (!owner) {
-        // Handle notification
-        return;
-      }
-      const tx = await contract.write.setMarketplace([marketplace]);
-
-      // Handle success
-      console.log("🚀 ~ onUpdateMarketplace ~ tx:", tx);
-      fetchConfig();
-      alert("Marketplace updated successfully");
-    } catch (error) {
-      // Handle error
-      console.log("🚀 ~ onUpdateMarketplace ~ error:", error);
-      handleError(error);
-    }
-  };
-
-  const onUpdateForwarder = async () => {
-    try {
-      const contract = handleControllerContract();
-      if (!contract) {
-        // Handle notification
-        return;
-      }
-      if (!owner) {
-        // Handle notification
-        return;
-      }
-      const tx = await contract.write.setForwarder([forwarder]);
-
-      // Handle success
-      console.log("🚀 ~ onUpdateForwarder ~ tx:", tx);
-      fetchConfig();
-      alert("Forwarder updated successfully");
-    } catch (error) {
-      // Handle error
-      console.log("🚀 ~ onUpdateForwarder ~ error:", error);
-      handleError(error);
-    }
-  };
-
-  const onUpdateVerifier = async () => {
-    try {
-      const contract = handleControllerContract();
-      if (!contract) {
-        // Handle notification
-        return;
-      }
-      if (!owner) {
-        // Handle notification
-        return;
-      }
-      const tx = await contract.write.setVerifier([verifier]);
+      const tx = await contract.write.setHandleTokenURIContract([handleTokenURI]);
 
       // Handle success
       console.log("🚀 ~ onUpdateVerifier ~ tx:", tx);
       fetchConfig();
-      alert("Verifier updated successfully");
+      alert("Handle Token URI updated successfully");
     } catch (error) {
       // Handle error
       console.log("🚀 ~ onUpdateVerifier ~ error:", error);
@@ -235,125 +131,69 @@ export default function HandleCard() {
 
   useEffect(() => {
     fetchConfig();
-  }, [walletClient]);
+  }, [walletClient, handleSelected]);
+
   return (
     <div className="card-admin">
       <a href="#">
-        <h5 className="card-title">Handle Controller</h5>
+        <h5 className="card-title">Handle Config</h5>
+        
       </a>
       <div className="mb-5">
-        <label className="form-label">Owner</label>
+        <label className="form-label">Dropdown Address</label>
+        <select
+          className="form-select"
+          value={handleSelected}
+          onChange={(e) => {
+            setHandleSelected(e.target.value as `0x${string}`);
+          }}
+        >
+          {handleAddresses.map((handle) => (
+            <option key={handle.address} value={getAddress(handle.address)}>
+              {handle.name}
+            </option>
+          ))}
+        </select>
+        <h3>Selected Handle Name: {name}</h3>
+        <h3>Selected Handle Symbol: {symbol}</h3>
+      </div>
+      <div className="mb-5">
+        <label className="form-label">Operator</label>
         <input
           className="form-input"
-          value={owner}
+          value={operator}
           onChange={(e) => {
-            setOwner(e.target.value);
+            setOperator(e.target.value);
           }}
         />
-        <button className="btn-danger" onClick={onUpdateOwner}>
-          Update owner
+        <button className="btn-danger" onClick={onUpdateOperator}>
+          Update operator
         </button>
       </div>
       <div className="mb-5">
-        <label className="form-label">Verifier</label>
+        <label className="form-label">Controller</label>
         <input
           className="form-input"
-          value={verifier}
+          value={controller}
           onChange={(e) => {
-            setVerifier(e.target.value);
+            setController(e.target.value);
           }}
         />
-        <button className="btn-danger" onClick={onUpdateVerifier}>
-          Update verifier
+        <button className="btn-danger" onClick={onUpdateController}>
+          Update controller
         </button>
       </div>
       <div className="mb-5">
-        <label className="form-label">Forwarder</label>
+        <label className="form-label">Handle Token URI</label>
         <input
           className="form-input"
-          value={forwarder}
+          value={handleTokenURI}
           onChange={(e) => {
-            setForwarder(e.target.value);
+            setHandleTokenURI(e.target.value);
           }}
         />
-        <button className="btn-danger" onClick={onUpdateForwarder}>
-          Update forwarder
-        </button>
-      </div>
-      <div className="mb-5">
-        <label className="form-label">Marketplace</label>
-        <input
-          className="form-input"
-          value={marketplace}
-          onChange={(e) => {
-            setMarketplace(e.target.value);
-          }}
-        />
-        <button className="btn-danger" onClick={onUpdateMarketplace}>
-          Update marketplace
-        </button>
-      </div>
-      <div className="mb-5">
-        <label className="form-label">Collector</label>
-        <input
-          className="form-input"
-          value={collector}
-          onChange={(e) => {
-            setCollector(e.target.value);
-          }}
-        />
-        <button className="btn-danger" onClick={onUpdateCollector}>
-          Update collector
-        </button>
-      </div>
-      <div className="mb-5">
-        <label className="form-label">Marketplace auction config</label>
-        <div className="grid grid-cols-5 grid-rows-5 gap-4">
-          <div className="col-span-5 row-span-3">
-            <label className="form-label">Buyout Bid Amount</label>
-            <input
-              type="number"
-              className="form-input"
-              value={auctionConfig?.buyoutBidAmount}
-              onChange={(e) => {
-                setAuctionConfig({
-                  ...auctionConfig,
-                  buyoutBidAmount: e.target.value,
-                });
-              }}
-            />
-          </div>
-          <div className="col-span-5 row-span-3">
-            <label className="form-label">Time Buffer In Seconds</label>
-            <input
-              type="number"
-              className="form-input"
-              value={auctionConfig?.timeBufferInSeconds}
-              onChange={(e) => {
-                setAuctionConfig({
-                  ...auctionConfig,
-                  timeBufferInSeconds: e.target.value,
-                });
-              }}
-            />
-          </div>
-          <div className="col-span-5 row-span-3">
-            <label className="form-label">Bid buffer Basic point</label>
-            <input
-              type="number"
-              className="form-input"
-              value={auctionConfig?.bidBufferBps}
-              onChange={(e) => {
-                setAuctionConfig({
-                  ...auctionConfig,
-                  bidBufferBps: e.target.value,
-                });
-              }}
-            />
-          </div>
-        </div>
-        <button className="btn-danger" onClick={onUpdateAuctionConfig}>
-          Update auction config
+        <button className="btn-danger" onClick={onUpdateHandleTokenURI}>
+          Update HandleTokenURI
         </button>
       </div>
     </div>
