+++ src/contract/contractConfig.ts
@@ -3,6 +3,7 @@ import BicRedeemFactory from './abi/BicRedeemFactory.json';
 import BicRedeemToken from './abi/BicRedeemToken.json';
 import BicTokenPaymaster from './abi/BicTokenPaymaster.json';
 import HandleController from './abi/HandleController.json';
+import Handle from './abi/Handle.json';
 import {createPublicClient, getContract} from "viem";
 import {sepolia, arbitrumSepolia} from "wagmi/chains";
 import {http} from "wagmi";
@@ -31,6 +32,12 @@ export const HandleControllerConfig = {
     address: process.env.NEXT_PUBLIC_HANDLE_CONTROLLER_ADDRESS as `0x${string}`,
     abi: HandleController
 }
+
+export const HandleConfig = (address: `0x${string}`) => ({
+    address: address as `0x${string}`,
+    abi: Handle
+})
+
 
 
 export const client = createPublicClient({
