+++ src/util/constant.ts
@@ -1,0 +1,27 @@
+export const handleAddresses = [
+    {
+        name: "Ownership name",
+        address: process.env.NEXT_PUBLIC_HANDLE_OWNERSHIP_USERNAME_ADDRESS as `0x${string}`
+    },
+    {
+        name: "Ownership personal",
+        address: process.env.NEXT_PUBLIC_HANDLE_OWNERSHIP_PERSONAL_NAME_ADDRESS as `0x${string}`
+    },
+    {
+        name: "Ownership community",
+        address: process.env.NEXT_PUBLIC_HANDLE_OWNERSHIP_COMMUNITY_NAME_ADDRESS as `0x${string}`
+    },
+
+    {
+        name: "BASE name",
+        address: process.env.NEXT_PUBLIC_HANDLE_BASE_USERNAME_ADDRESS as `0x${string}`
+    },
+    {
+        name: "BASE personal",
+        address: process.env.NEXT_PUBLIC_HANDLE_BASE_PERSONAL_NAME_ADDRESS as `0x${string}`
+    },
+    {
+        name: "BASE community",
+        address: process.env.NEXT_PUBLIC_HANDLE_BASE_COMMUNITY_NAME_ADDRESS as `0x${string}`
+    },
+]
